requirements.txt
apache-airflow
tweepy
pytest